requirements.txt
requests
responses
pytest
python-dotenv